Trait scad_compiler::parsing::Parsing

pub trait Parsing: Database + HasQueryGroup<ParsingStorage> {
    fn src(&self) -> Text;
    fn set_src(&mut self, value__: Text);
    fn set_src_with_durability(&mut self, value__: Text, durability__: Durability);
    fn parse(&self, key0: Text) -> (Package, Diagnostics);
}

Required Methods

fn src(&self) -> Text

The program’s source code.

fn set_src(&mut self, value__: Text)

Set the value of the src input.

See src for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_src_with_durability(&mut self, value__: Text, durability__: Durability)

Set the value of the src input and promise that its value will never change again.

See src for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn parse(&self, key0: Text) -> (Package, Diagnostics)

Parse some source code into an AST using the scad_syntax crate.